work/135/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
geoip2==4.7.0
user-agents==2.2.0
python-dateutil==2.8.2
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
elasticsearch==8.11.0
aiofiles==23.2.1
anyio==3.7.1
qrcode==7.4.2
pillow==10.2.0
cachetools==5.3.2
